fix for damage patch
