+++ CSharpSourceCode/Battle/Damage/DamageType.cs
@@ -8,7 +8,6 @@ namespace TOW_Core.Battle.Damage
 {
     public enum DamageType
     {
-        Invalid,
         Physical,
         Magical,
         Fire,
